- added throwing of DataRetrievalFailureException if removed AnimalEnvironment is not persistent - added test of DataAccessException throwing if removing non-peristent animal

+++ src/main/java/cz/muni/fi/pa165/dominatingspecies/dao/AnimalEnvironmentDaoImpl.java
@@ -4,6 +4,7 @@ import cz.muni.fi.pa165.dominatingspecies.entity.AnimalEnvironment;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import org.springframework.dao.DataRetrievalFailureException;
 import org.springframework.stereotype.Repository;
 
 /**
@@ -32,6 +33,10 @@ public class AnimalEnvironmentDaoImpl implements AnimalEnvironmentDao {
 
     @Override
     public void remove(AnimalEnvironment environment) {
+        if (!em.contains(environment)) {
+            throw new DataRetrievalFailureException("Entity " + environment + " does not exist in the persistent storage");
+        }
+        
         em.remove(environment);
     }
 }
